Add the implementation of matmul_v4.cu
The implementation is based on matmul_v1.cu. matmul_v4.cu adds the
the implementation of thread tile, thus the thread mapping and tile size
were completely changed.

The TFLOPS is 12.12 / 18.38 = 65.94% of cuBLAS, the grid dim is (
(32, 32), the block dim is (16, 16), the block tile dim is (128, 8)
for tileA, tileB, (128, 128) for tileC, the thread tile dim is (8, 8)