Add the implementation of matmul_v6/v7.cu
The implementation of matmul_v6.cu is based on matmul_v5.cu. The only
difference between matmul_v5.cu and matmul_v6.cu is the value of
BLOCK_TILE_K(from 8 to 16). Larger BLOCK_TILE_K is advantageous for
further reduce global memory accesses. The differences between
matmul_v6.cu and matmul_v7.cu is that matmul_v7.cu implements a 2 stage
pipeline, which further improves TFLOPS.

The TFLOPS of matmul_v6 is 16.58 / 18.38 = 90.20% of cuBLAS, the TFLOPS
of matmul_v7 is 16.72 / 18.38 = 90.96% of cuBLAS. For both version, the
grid dim is (32, 32), the block dim is (16, 16), the block tile dim is
(128, 16) for tileA, tileB, (128, 128) for tileC, and the thread tile
dim is (8, 8).