Add the implementation of matmul_t_v1.cu
matmul_t_v1.cu is based on matmul_t_v0.cu, and it uses a padding
startegy to solve the shared memory access bank confilt problems in
matmul_t_v0.cu.

The TFLOPS is 99.54 / 222.11 = 44.81% of cuBLAS. The blockdim is
(16, 16). The block tile shape is (256, 128) for tileC, (256, 16) for
tileA and (16, 128) for tileB. The warp tile shape is
(64, 64), and the grid dim is (16, 32).