Add the implementation of matmul_t_v3.cu
matmul_t_v3.cu is based on matmul_t_v2.cu, and it uses a swizzle trick
to avoid global to shared store and shared to register load bank
conflicts. The performance improvement relative to matmul_t_v2.cu is
relatively low, because the shared memory usage is too high so that
the occupancy is extremely low.

The TFLOPS is 128.11 / 222.11 = 57.67% of cuBLAS. The blockdim is
(16, 16). The block tile shape is (256, 128) for tileC, (256, 32) for
tileA and (32, 128) for tileB. The warp tile shape is
(64, 64), and the grid dim is (16, 32).