Add the implementation of matmul_t_v4.cu
matmul_t_v4.cu is based on matmul_t_v3.cu, and it uses a swizzle trick
to avoid global to shared store and shared to register load bank
conflicts. This implementation change BLOCK_TILE_Y from 256 to 128 to
reduce shared memory usage, but the occupancy improvement is relatively
low.

The TFLOPS is 152.64 / 255.31 = 59.78% of cuBLAS. The blockdim is
(128, 1). The block tile shape is (128, 128) for tileC, (128, 32) for
tileA and (32, 128) for tileB. The warp tile shape is
(64, 64), and the grid dim is (32, 32).